requirements.txt
# Django and core dependencies
Django>=5.2.4,<5.3
djangorestframework>=3.14.0,<3.15

# Firebase integration
firebase-admin>=6.2.0,<6.3

# Development and testing
pytest>=7.4.0,<7.5
pytest-django>=4.5.0,<4.6
black>=23.0.0,<24.0
flake8>=6.0.0,<6.1

# Database (SQLite is included with Django)
# For production, consider:
 psycopg2-binary>=2.9.0,<2.10  # PostgreSQL
# mysqlclient>=2.1.0,<2.2       # MySQL

# Environment management
python-dotenv>=1.0.0,<1.1

# CORS (if needed for frontend integration)
django-cors-headers>=4.0.0,<4.1














